agroguard/ml-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
tensorflow==2.15.0
numpy==1.26.3
pillow==10.2.0
scikit-learn==1.4.2
